+++ NEWS.md
@@ -5,7 +5,11 @@
 - All function arguments which expected a list of quosures created by `vars()`
 are now expecting a list of expressions created by `exprs()`. For example,
 instead of `by_vars = vars(STUDYID, USUBJID)` `by_vars = exprs(STUDYID,
-USUBJID)` must be used now.
+USUBJID)` must be used now. (#197)
+
+## Documentation
+
+- New vignette "Create `ADTR`" (#16)
 
 ## Various